test: simplify asset test organization
